create POST /api/pictures with no validation
